+++ cloudsim-plus/src/main/java/org/cloudbus/cloudsim/allocationpolicies/VmAllocationPolicySimple.java
@@ -67,7 +67,7 @@ public class VmAllocationPolicySimple extends VmAllocationPolicyAbstract {
             final int hostFreePes = entry.getValue();
             if (host.vmCreate(vm)) {
                 mapVmToPm(vm, host);
-                getUsedPes().put(vm, vm.getNumberOfPes());
+                addUsedPes(vm);
                 getHostFreePesMap().put(host, hostFreePes - vm.getNumberOfPes());
                 if(!hostsWhereVmCreationFailed.isEmpty()){
                     Log.printFormattedLine("[VmAllocationPolicy] VM #%d was successfully allocated to Host #%d", vm.getId(), host.getId());
@@ -89,7 +89,7 @@ public class VmAllocationPolicySimple extends VmAllocationPolicyAbstract {
 
         mapVmToPm(vm, host);
         final int requiredPes = vm.getNumberOfPes();
-        getUsedPes().put(vm, requiredPes);
+        addUsedPes(vm);
         getHostFreePesMap().put(host, getHostFreePesMap().get(host) - requiredPes);
 
         Log.printFormattedLine(
@@ -121,7 +121,7 @@ public class VmAllocationPolicySimple extends VmAllocationPolicyAbstract {
     @Override
     public void deallocateHostForVm(Vm vm) {
         Host host = unmapVmFromPm(vm);
-        int pes = getUsedPes().remove(vm);
+        int pes = removeUsedPes(vm);
         if (host != Host.NULL) {
             host.destroyVm(vm);
             getHostFreePesMap().put(host, getHostFreePesMap().get(host) + pes);